+++ core-ktx/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Unreleased
 
+### Dependencies
+
+- androidx.core 1.3.2 -> 1.5.0
+
+### Added
+
+- `View.isKeyboardVisible: Boolean` and `Activity.isKeyboardVisible: Boolean`
+- `View.showKeyboard()`
+- `View.hideKeyboard()` and `Activity.hideKeyboard()`
+
 ## 1.3.2-0
 
 First release